process handling: PROCTABLE: separate load/store from selection
