+++ libpromises/processes_select_sysprobe.c
@@ -37,19 +37,9 @@
 #include <process_unix_priv.h>
 #include <process_lib.h>
 
-#include <string.h>
-#include <sys/user.h>
-
 #include <eval_context.h>
-#include <files_names.h>
 #include <conversion.h>
-#include <item_lib.h>
-#include <dir.h>
 
-
-#define NPROC_GUESS 500
-
-static JsonElement *PROCTABLE = NULL;
 
 /*
  * internal functions
@@ -247,7 +237,9 @@ Item *SelectProcesses(const char *process_name, const ProcessSelect *a, bool att
     assert(a != NULL);
     Item *result = NULL;
 
-    if (PROCTABLE == NULL)
+    const JsonElement *ptable;
+    ptable = FetchProcessTable();
+    if (ptable == NULL)
     {
         Log(LOG_LEVEL_ERR, "%s: PROCESSTABLE is empty", __func__);
         return NULL;
@@ -256,7 +248,7 @@ Item *SelectProcesses(const char *process_name, const ProcessSelect *a, bool att
     const char *cmd;
     pid_t pid;
     const JsonElement *pdata;
-    JsonIterator iter = JsonIteratorInit(PROCTABLE);
+    JsonIterator iter = JsonIteratorInit(ptable);
     while ((pdata = JsonIteratorNextValue(&iter)))
     {
         pid = IntFromString(JsonIteratorCurrentKey(&iter));
@@ -322,7 +314,9 @@ bool IsProcessNameRunning(char *procNameRegex)
 {
     bool matched = false;
 
-    if (PROCTABLE == NULL)
+    const JsonElement *ptable;
+    ptable = FetchProcessTable();
+    if (ptable == NULL)
     {
         Log(LOG_LEVEL_ERR, "%s: PROCESSTABLE is empty", __func__);
         return false;
@@ -330,7 +324,7 @@ bool IsProcessNameRunning(char *procNameRegex)
 
     const JsonElement *pdata;
     const char *cmd;
-    JsonIterator iter = JsonIteratorInit(PROCTABLE);
+    JsonIterator iter = JsonIteratorInit(ptable);
     while (!matched && (pdata = JsonIteratorNextValue(&iter)))
     {
             cmd = JsonObjectGetAsString(pdata, JPROC_KEY_CMD);
@@ -343,176 +337,4 @@ bool IsProcessNameRunning(char *procNameRegex)
     }
 
     return matched;
-}
-
-
-const char *GetProcessTableLegend(void)
-{
-    if (PROCTABLE)
-    {
-        return "<via /proc>";
-    }
-    else
-    {
-        return "<Process table not loaded>";
-    }
-}
-
-
-/*
- * LoadProcessTable() and subordinates
- *
- * ClearProcessTable()
- */
-
-/* "/proc/nnn" is a process directory only if pure integer "nnn" */
-static bool IsProcDir(const char *name)
-{
-    const char *p = name;
-
-    while (*p)
-    {
-        if (!isdigit(*p)) {
-            return false;
-        }
-        p++;
-    }
-
-    return true;
-}
-
-
-bool LoadProcessTable()
-{
-    Dir *dirh = NULL;
-    const struct dirent *dirp;
-
-    if (PROCTABLE)
-    {
-        Log(LOG_LEVEL_VERBOSE, "Reusing cached process table");
-        return true;
-    }
-
-    if ((dirh = DirOpen(PROCDIR)) == NULL)
-    {
-        Log(LOG_LEVEL_ERR, "Unable to open %s directory'. (opendir: %s)", PROCDIR, GetErrorStr());
-        return false;
-    }
-
-    PROCTABLE = JsonObjectCreate(NPROC_GUESS);
-
-    pid_t  pid;
-    JsonElement *pdata;
-    while ((dirp = DirRead(dirh)) != NULL)
-    {
-        /*
-         * Process next entry. Skip non-numeric names as being non-process.
-         */
-        if (! IsProcDir(dirp->d_name))
-        {
-            continue;
-        }
-
-        pid = atol(dirp->d_name);
-
-        /* It ought to be a directory... */
-        if (dirp->d_type != DT_DIR)
-        {
-            Log(LOG_LEVEL_ERR, "'%s/%s' not a directory\n", PROCDIR, dirp->d_name);
-            continue;
-        }
-
-        pdata = LoadProcStat(pid);
-        if (pdata == NULL)
-        {
-            Log(LOG_LEVEL_ERR, "failure creating 'stat' data for '%s'\n", dirp->d_name);
-            continue;
-        }
-
-        JsonObjectAppendObject(PROCTABLE, dirp->d_name, pdata);
-
-    }
-
-    DirClose(dirh);
-
-    return true;
-
-}
-
-/*
- * return a read-only pointer to the JSON version of the process table
- */
-const JsonElement *FetchProcessTable()
-{
-    if (! LoadProcessTable())
-    {
-        Log(LOG_LEVEL_ERR, "Unable to load process table");
-        return NULL;
-    }
-
-    return PROCTABLE;
-}
-
-void ClearProcessTable(void)
-{
-    JsonDestroy(PROCTABLE);
-    PROCTABLE= NULL;
-}
-
-
-/*
- * Access points for individual processes.
- * Independent of complete PROCTABLE.
- * Initial use is mostly unit tests
- * although there are a couple of other uses.
- */
-
-time_t GetProcessStartTime(pid_t pid)
-{
-    JsonElement *pdata;
-
-    pdata = LoadProcStat(pid);
-    if (pdata) {
-        time_t t = IntFromString(JsonObjectGetAsString(pdata, JPROC_KEY_STARTTIME_BOOT));
-
-        JsonDestroy(pdata);
-
-        return t;
-    }
-    else
-    {
-        return PROCESS_START_TIME_UNKNOWN;
-    }
-}
-
-ProcessState GetProcessState(pid_t pid)
-{
-    JsonElement *pdata;
-
-    pdata = LoadProcStat(pid);
-    if (pdata) {
-        ProcessState pstate;
-
-        const char *status = JsonObjectGetAsString(pdata, JPROC_KEY_PSTATE);
-        switch (status[0])
-        {
-        case 'T':
-            pstate = PROCESS_STATE_STOPPED;
-            break;
-        case 'Z':
-            pstate = PROCESS_STATE_ZOMBIE;
-            break;
-        default:
-            pstate = PROCESS_STATE_RUNNING;
-            break;
-        }
-
-        JsonDestroy(pdata);
-
-        return pstate;
-    }
-    else
-    {
-        return PROCESS_STATE_DOES_NOT_EXIST;
-    }
 }
